+++ application/views/identifiers/list.php
@@ -3,7 +3,12 @@ defined('BASEPATH') or exit('No direct script access allowed');
 ob_start();
 ?>
 
-<h1 class="mt-5">Liste des types d'identifiants <a class="btn btn-outline-primary" href="/identifiers/new" role="button">Ajouter</a></h1>
+<h1 class="mt-5">
+  Liste des types d'identifiants
+  <?php if ($controller->hasPermission('project_identifiers', 'add')): ?>
+    <a class="btn btn-outline-primary" href="/identifiers/new" role="button">Ajouter</a>
+  <?php endif; ?>
+</h1>
 <p class="lead">Page listant les différents types d'identifiants</p>
 
 <table class="table table-hover">
@@ -18,9 +23,15 @@ ob_start();
     <tr class="searcher-item" data-searcher="<?php echo strtolower(htmlspecialchars($identifier->name)); ?>">
       <td><?php echo $identifier->name; ?></td>
       <td>
-        <a href="/identifier/edit/<?php echo $identifier->id; ?>">Modifier</a>
-        -
-        <a href="/identifier/delete/<?php echo $identifier->id; ?>">Supprimer</a>
+        <?php if ($controller->hasPermission('project_identifiers', 'edit')): ?>
+          <a href="/identifier/edit/<?php echo $identifier->id; ?>">Modifier</a>
+        <?php endif; ?>
+        <?php if ($controller->hasPermission('project_identifiers', 'edit') && $controller->hasPermission('project_identifiers', 'delete')): ?>
+          -
+        <?php endif; ?>
+        <?php if ($controller->hasPermission('project_identifiers', 'delete')): ?>
+          <a href="/identifier/delete/<?php echo $identifier->id; ?>">Supprimer</a>
+        <?php endif; ?>
       </td>
     </tr>
     <?php endforeach; ?>
